refactor: show overnight upload in backup controller
